Move camera_controller to controllers/ folder
